+++ packages/component-registry/src/index.ts
@@ -1,5 +1,6 @@
 export * from './types.js';
 export * from './registry.js';
+export * from './labels.js';
 export * from './spec.js';
 export * from './schema.js';
 export * from './factory.js';